requirements.txt
asgiref==3.2.7
Babel==2.8.0
dj-database-url==0.5.0
Django==3.0.7
django-crispy-forms==1.9.1
django-filter==2.3.0
django-jazzmin==2.1.1
django-phonenumber-field==4.0.0
djangorestframework==3.11.0
gunicorn==20.0.4
importlib-metadata==1.6.1
Markdown==3.2.2
phonenumbers==8.12.5
Pillow==7.1.2
psycopg2==2.8.5
pytz==2020.1
sqlparse==0.3.1
validate-email==1.3
whitenoise==5.1.0
zipp==3.1.0
